+++ asset/src/asset-helpers.cpp
@@ -31,7 +31,7 @@ AssetExpected<uint32_t> checkElementIdentifier(const std::string& paramName, con
     }
 
     if (auto eid = db::nameToAssetId(paramValue)) {
-        return uint32_t(*eid);
+        return *eid;
     } else {
         std::string err      = "value '{}' is not valid identifier. Error: {}"_tr.format(paramValue, eid.error());
         std::string expected = "existing identifier"_tr;
